// src/types/api.d.ts
export interface components {
    schemas: {
        Client: {
            razonSocial: string;
            businessType: string;
            name: string;
            fiscalAddress: string;
            email: string;
            clientLocations?: components["schemas"]["ClientLocation"][] | null;
            phoneNumber: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        Client2: {
            razonSocial: string;
            businessType: string;
            name: string;
            fiscalAddress: string;
            email: string;
            clientLocations?: components["schemas"];
            phoneNumber: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        Client3: {
            razonSocial: string;
            businessType: string;
            name: string;
            fiscalAddress: string;
            email: string;
            clientLocations?: components["schemas"]["ClientLocation2"][] | null;
            phoneNumber: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        Client4: {
            razonSocial: string;
            businessType: string;
            name: string;
            fiscalAddress: string;
            email: string;
            clientLocations?: components["schemas"];
            phoneNumber: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        ClientLocation: {
            client?: components["schemas"]["Client2"];
            address: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        ClientLocation2: {
            client?: components["schemas"]["Client4"];
            address: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        ClientPatchDTO: {
            razonSocial: string | null;
            businessType: string | null;
            name: string | null;
            fiscalAddress: string | null;
            email: string | null;
            phoneNumber: string | null;
        };
        LoginRequest: {
            /** @default admin@admin.com */
            email: string;
            /** @default Acide2025/1 */
            password: string;
        };
        LoginResponse: {
            accessToken: string;
            refreshToken: string;
            /** Format: int32 */
            accessExpiresIn: number;
            /** Format: int32 */
            refreshExpiresIn: number;
        };
        ProblemDetails: {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
        };
        TermsAndConditions: {
            name: string;
            content: string;
            /** Format: uuid */
            id?: string;
            isActive?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            modifiedAt?: string;
        };
        TermsAndConditionsPatchDTO: {
            name?: string | null;
            content?: string | null;
        };
        UnauthorizedResult: {
            /** Format: int32 */
            statusCode?: number;
        };
        UserReturn: {
            username: string;
            email: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}